+++ src/create_file_system.c
@@ -1,9 +1,42 @@
 #include<stdio.h>
 #include<stdlib.h>
 #include<string.h>
-#include "../header/create_file_system.h"
 #include "../header/Global.h"
 #include "../header/I_O_operation.h"
+
+//-------------------------------------------------------------------Variables-------------------------------------------------------
+
+// Defining the MetaData Variables
+
+// Size of sizeHDD
+unsigned long long int sizeHDD;
+
+// Size of Start Block
+unsigned long long int size_StartBlock;
+
+// HDD - Size of Start_Block
+unsigned long long int sizeHDD_StartBlock;
+
+// Size of the MetaData
+unsigned long long int size_MetaData;
+
+// Size of the DiskBlock
+unsigned long long int size_DiskBlock;
+
+// MetaData + DiskBlock
+unsigned long long int size_MetaData_DiskBlock;
+
+// Number of MetaData
+unsigned long long int num_MetaData;
+
+// Number of the DiskBlock
+unsigned long long int num_DiskBlock;
+
+unsigned long long int Space;
+
+//-------------------------------------------------------------------Declaring the Structures-------------------------------------------------------
+
+struct StartBlock SB;
 
 // Creating the File System
 void create_file_system() {
