+++ src/main.c
@@ -1,19 +1,17 @@
 #include<stdio.h>
 #include<stdlib.h>
 #include "../header/create_file_system.h"
-#include "../header/Global.h"
-#include "../header/I_O_operation.h"
 
 
 int main() {
     
     create_file_system();
     read_file_system(1);
-    format_file_system();
-    read_file_system(1);
-    closing_file_system(1);
-    mount_file_system(1);
-    read_file_system(1);
-    closing_file_system();
+    // format_file_system();
+    // read_file_system(1);
+    // closing_file_system(1);
+    // mount_file_system(1);
+    // read_file_system(1);
+    // closing_file_system();
     return 0;
 }
